Behavior update
Creatures no long loop within the environment. If they reach the edge of
the screen, they are stuck there. To combat this, the creatures now have
inputs of where they are in the world.

Vision is updated to one eye instead of two. There are 10 'degree slots'
for their vision. When the closet food is found within their field of
view, it calculates the angle to that food where 0 is on the far left of
their vision and 1 is on the far right of their vision. This means if
food is straight ahead, the 'angle' will be 1/2. Of the ten 'angle
slots', the two closest to the angle towards the food will be given the
distance to the food (0 is far away, 1 is close). Nearby angle-slots
will be given a portion of that distance.

A red line is drawn from the creature towards the food it is closest to.